+++ playground/vite-nitro/server/api/ssr/radiant-component-server-card.ts
@@ -3,9 +3,5 @@ import { defineHandler } from 'nitro';
 export default defineHandler(async () => {
 	const { renderSsrComponent, createSsrResponse } = await import('../../render');
 	const { RadiantComponentServerCardElement } = await import('@/components/radiant-component-server-card.script');
-	return createSsrResponse(
-		await renderSsrComponent({
-			component: RadiantComponentServerCardElement,
-		}),
-	);
+	return createSsrResponse(await renderSsrComponent(RadiantComponentServerCardElement));
 });